Restructured widget object to be stored in API object
